Prevent MainActivity from becoming a god class
